// src/app/api/models/reason-enum.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * * `DAMAGE` - Damage
 * * `LOSS` - Loss
 * * `SHOP_USE` - Shop use
 * * `WARRANTY_REPLACEMENT` - Warranty replacement
 * * `RETURN_TO_SUPPLIER` - Return to supplier
 * * `SUPPLIER_REPLACEMENT` - Supplier replacement
 * * `OPENING_BALANCE` - Opening balance
 */
export type ReasonEnum = 'DAMAGE' | 'LOSS' | 'SHOP_USE' | 'WARRANTY_REPLACEMENT' | 'RETURN_TO_SUPPLIER' | 'SUPPLIER_REPLACEMENT' | 'OPENING_BALANCE';